requirements.txt
absl-py==0.8.1
nibabel==3.0.0
numpy==1.18.2
pandas==0.24.2
scikit-image==0.16.2
tensorflow==1.13.1
